FDK. VariableFontSupport. Fix some Windows compile issue, fix crashing stack count bug in t2charstring.c

+++ FDK/Tools/Programs/public/lib/api/absfont.h
@@ -245,7 +245,7 @@ typedef struct
     long LanguageGroup;         /* 0 */
     float ExpansionFactor;      /* 0.06 */
     float initialRandomSeed;    /* 0. */
-    long vsindex;
+    unsigned short vsindex;
     struct
     {
         struct
